refactor: rename getExchangeDescription function import and update location text logic in BookOwnerCard
