+++ src/widgets/book-owner-card/lib/get-exchange-description.ts
